[project @ 2004-09-23 17:30:32 by nick] Fixes and restructuring for bug 554:
1. need to talk about "appropriate privilege", not "root"

2. NIS/YP is an undefined term (and in other places, it is just plain "NIS").
Use the term "NIS" (and consider adding it to the terminology chapter).

3. " When called with one argument or with the --file option, the commands set
the host name or the NIS/YP domain name. Note, that only the super-user can
change the names." combines both the above problems AND suggests that "hostname
-d" should set the host name to "-d". The synopsis also forbids setting the NIS
domain this way. It is not an "argument" ... it is an "operand".

4. There is confusion between the NIS domain and the DNS domain in places.